+++ test/access-token.js
@@ -14,17 +14,12 @@ const AccessToken = require('../lib/access-token');
 const { Client } = require('../lib/client');
 const { has, hasIn } = require('./_property');
 const { createModuleConfigWithDefaults: createModuleConfig } = require('./_module-config');
-const { createAuthorizationServer } = require('./_authorization-server-mock');
+const { createAuthorizationServer, getHeaderCredentialsScopeOptions } = require('./_authorization-server-mock');
 
 const chance = new Chance();
 chance.mixin({ accessToken: accessTokenMixin });
 
-const scopeOptions = {
-  reqheaders: {
-    Accept: 'application/json',
-    Authorization: 'Basic dGhlK2NsaWVudCtpZDp0aGUrY2xpZW50K3NlY3JldA==',
-  },
-};
+const scopeOptions = getHeaderCredentialsScopeOptions();
 
 test('@create => throws an error when no token payload is provided', (t) => {
   const config = createModuleConfig();
@@ -347,7 +342,6 @@ test.serial('@refresh => creates a new access token with a custom token path', a
 
   const refreshParams = {
     grant_type: 'refresh_token',
-    scope: 'TESTING_EXAMPLE_SCOPES',
     refresh_token: accessTokenResponse.refresh_token,
   };
 
@@ -355,7 +349,73 @@ test.serial('@refresh => creates a new access token with a custom token path', a
   const scope = server.tokenSuccessWithCustomPath('/the-custom/path', scopeOptions, refreshParams);
 
   const accessToken = new AccessToken(config, client, accessTokenResponse);
-  const refreshAccessToken = await accessToken.refresh({ scope: 'TESTING_EXAMPLE_SCOPES' });
+  const refreshAccessToken = await accessToken.refresh();
+
+  scope.done();
+  t.true(has(refreshAccessToken.token, 'access_token'));
+});
+
+test.serial('@refresh => creates a new access token with custom (inline) http options', async (t) => {
+  const config = createModuleConfig();
+
+  const accessTokenResponse = chance.accessToken({
+    expireMode: 'expires_in',
+  });
+
+  const client = new Client(config);
+
+  const refreshParams = {
+    grant_type: 'refresh_token',
+    refresh_token: accessTokenResponse.refresh_token,
+  };
+
+  const customScopeOptions = getHeaderCredentialsScopeOptions({
+    reqheaders: {
+      'X-REQUEST-ID': 123,
+    },
+  });
+
+  const server = createAuthorizationServer('https://authorization-server.org:443');
+  const scope = server.tokenSuccess(customScopeOptions, refreshParams);
+
+  const httpOptions = {
+    headers: {
+      'X-REQUEST-ID': 123,
+    },
+  };
+
+  const accessToken = new AccessToken(config, client, accessTokenResponse);
+  const refreshAccessToken = await accessToken.refresh(null, httpOptions);
+
+  scope.done();
+  t.true(has(refreshAccessToken.token, 'access_token'));
+});
+
+test.serial('@refresh => creates a new access token with custom (inline) http options without overriding (required) http options', async (t) => {
+  const config = createModuleConfig();
+
+  const accessTokenResponse = chance.accessToken({
+    expireMode: 'expires_in',
+  });
+
+  const client = new Client(config);
+
+  const refreshParams = {
+    grant_type: 'refresh_token',
+    refresh_token: accessTokenResponse.refresh_token,
+  };
+
+  const server = createAuthorizationServer('https://authorization-server.org:443');
+  const scope = server.tokenSuccess(scopeOptions, refreshParams);
+
+  const httpOptions = {
+    headers: {
+      Authorization: 'Basic credentials',
+    },
+  };
+
+  const accessToken = new AccessToken(config, client, accessTokenResponse);
+  const refreshAccessToken = await accessToken.refresh(null, httpOptions);
 
   scope.done();
   t.true(has(refreshAccessToken.token, 'access_token'));
@@ -435,6 +495,70 @@ test.serial('@revoke => performs a token revoke with a custom revoke path', asyn
   scope.done();
 });
 
+test.serial('@revoke => performs a token revoke with custom (inline) http options', async (t) => {
+  const config = createModuleConfig();
+  const client = new Client(config);
+
+  const accessTokenResponse = chance.accessToken({
+    expireMode: 'expires_in',
+  });
+
+  const revokeParams = {
+    token: accessTokenResponse.refresh_token,
+    token_type_hint: 'refresh_token',
+  };
+
+  const customScopeOptions = getHeaderCredentialsScopeOptions({
+    reqheaders: {
+      'X-REQUEST-ID': 123,
+    },
+  });
+
+  const server = createAuthorizationServer('https://authorization-server.org:443');
+  const scope = server.tokenRevokeSuccess(customScopeOptions, revokeParams);
+
+  const httpOptions = {
+    headers: {
+      'X-REQUEST-ID': 123,
+    },
+  };
+
+  const accessToken = new AccessToken(config, client, accessTokenResponse);
+
+  await t.notThrowsAsync(() => accessToken.revoke('refresh_token', httpOptions));
+
+  scope.done();
+});
+
+test.serial('@revoke => performs a token revoke with custom (inline) http options without overriding (required) http options', async (t) => {
+  const config = createModuleConfig();
+  const client = new Client(config);
+
+  const accessTokenResponse = chance.accessToken({
+    expireMode: 'expires_in',
+  });
+
+  const revokeParams = {
+    token: accessTokenResponse.refresh_token,
+    token_type_hint: 'refresh_token',
+  };
+
+  const server = createAuthorizationServer('https://authorization-server.org:443');
+  const scope = server.tokenRevokeSuccess(scopeOptions, revokeParams);
+
+  const httpOptions = {
+    headers: {
+      Authorization: 'Basic credentials',
+    },
+  };
+
+  const accessToken = new AccessToken(config, client, accessTokenResponse);
+
+  await t.notThrowsAsync(() => accessToken.revoke('refresh_token', httpOptions));
+
+  scope.done();
+});
+
 test.serial('@revoke => throws an error with an invalid tokenType option', async (t) => {
   const config = createModuleConfig();
   const client = new Client(config);
@@ -471,6 +595,80 @@ test.serial('@revokeAll => revokes both the access and refresh tokens', async (t
   const accessToken = new AccessToken(config, client, accessTokenResponse);
 
   await t.notThrowsAsync(() => accessToken.revokeAll());
+
+  scope.done();
+});
+
+test.serial('@revokenAll => revokes both the access and refresh tokens with custom (inline) http options', async (t) => {
+  const config = createModuleConfig();
+  const client = new Client(config);
+
+  const accessTokenResponse = chance.accessToken({
+    expireMode: 'expires_in',
+  });
+
+  const refreshTokenRevokeParams = {
+    token: accessTokenResponse.refresh_token,
+    token_type_hint: 'refresh_token',
+  };
+
+  const accessTokenRevokeParams = {
+    token: accessTokenResponse.access_token,
+    token_type_hint: 'access_token',
+  };
+
+  const customScopeOptions = getHeaderCredentialsScopeOptions({
+    reqheaders: {
+      'X-REQUEST-ID': 123,
+    },
+  });
+
+  const server = createAuthorizationServer('https://authorization-server.org:443');
+  const scope = server.tokenRevokeAllSuccess(customScopeOptions, accessTokenRevokeParams, refreshTokenRevokeParams);
+
+  const httpOptions = {
+    headers: {
+      'X-REQUEST-ID': 123,
+    },
+  };
+
+  const accessToken = new AccessToken(config, client, accessTokenResponse);
+
+  await t.notThrowsAsync(() => accessToken.revokeAll(httpOptions));
+
+  scope.done();
+});
+
+test.serial('@revokeAll => revokes tboth the access and refresh tokens with custom (inline) http options without overriding (required) http options', async (t) => {
+  const config = createModuleConfig();
+  const client = new Client(config);
+
+  const accessTokenResponse = chance.accessToken({
+    expireMode: 'expires_in',
+  });
+
+  const refreshTokenRevokeParams = {
+    token: accessTokenResponse.refresh_token,
+    token_type_hint: 'refresh_token',
+  };
+
+  const accessTokenRevokeParams = {
+    token: accessTokenResponse.access_token,
+    token_type_hint: 'access_token',
+  };
+
+  const server = createAuthorizationServer('https://authorization-server.org:443');
+  const scope = server.tokenRevokeAllSuccess(scopeOptions, accessTokenRevokeParams, refreshTokenRevokeParams);
+
+  const httpOptions = {
+    headers: {
+      Authorization: 'Basic credentials',
+    },
+  };
+
+  const accessToken = new AccessToken(config, client, accessTokenResponse);
+
+  await t.notThrowsAsync(() => accessToken.revokeAll(httpOptions));
 
   scope.done();
 });